requirements.txt
pymongo[srv]==4.0.1
python-dotenv==0.19.2
Flask~=2.0.3